48200 rows × 48 columns
The dataset named “48_Industry_Portfolios_daily.CSV”  consists of 48,200 records of daily returns for 48 industry portfolios from July 1st, 1926 to Oct 31st, 2017. Specifically, an industry portfolio is composed of a New York Stock Exchange (NYSE) stock, an American Stock Exchange (AMEX) stock and a Nasdaq (NASDAQ) stock from the same industry. In the original dataset, the index of the rows is the date, and the index of the columns is the name of the industry.